Update texas_county_data.csv
I've updated the CSV of county data to include a column for whether there is a PDO, I've updated a few other Odyssey links, and I've added notes for how far certain places are from Austin to try and identify places within 3.5 hours drive (at max). The date the Odyssey link was updated is also a column on the table.