Adds a hash of the underlying HTML to parser
For determining versions of files using hashes, I recommend we add a hash of the underlying HTML to the JSON file for the case as a field. It is otherwise not a step done anywhere else in this repo so far. This will later be used for creating an entry ID in CosmosDB.